FOLIO-2213: Remove EBSCO kubernetes descriptions
The outdated alternative-install/kubernetes-rancher/EBSCO
directory has been removed from the master branch of the
https://github.com/folio-org/folio-install repository
because it was unmaintained and had security issues:
https://github.com/folio-org/folio-install/pull/58

Therefore most of the "Kubernetes example" installation on
https://docs.folio.org/docs/getting-started/installation/kubernetesex/
can no longer point to those files and needs to be removed.

A link to the TAMU kubernetes descriptions
https://github.com/folio-org/folio-install/tree/master/alternative-install/kubernetes-rancher/TAMU
can be used instead.