Delete singleservernocontainers installation doc
The single server no containers installation document
is outdated, it's from November 2020.

It requires the Debian repository http://folio-repo.bib-bvb.de
this has been offline for a long time.

The Debian containers doesn't allow two tenants running
two different versions of the same module.

Therefore this installation document should be removed.

It is not lost because it remains in the git history:
https://github.com/folio-org/docs/commits/master