Document fresh install process
This rewrites the Single Server installation instructions to explain the process for a fresh install.  The instructions for Juniper and prior did assume a fresh install, but they were rewritten in Kiwi to be an upgrade from Juniper.  Those "upgrade" installations are useful for its own use case, but they are difficult to use in context of a fresh installation, because they require you to perform the Juniper installation first and then repeat or revise several of those steps for the current release.  Ingolf Kuss and I rewrote these.