Add clarification text for MG version of doc
Clarifying text so that there is one place you can link to in the check out docs that references both check out steps. (It's referenced at the very top, but if you want to jump to a specific place and not read the entire document, there is no place that mentions both steps - user scan, item scan. So this adds that.)