DOCS-220: docs.folio.org loses release menu and search box when browser gets resized
When the screen width is small or the text size is increased (browser settings or
browser zoom with Strg +) docsy moves elements like search box, language selector
and release menu from the navigation bar at the top into the menu bar at the left
or into the drow-down menu.

However, it currently doesn't work. Changes to sidebar-tree.html are needed to
fix the incorrect configuration, and changes to _sidebar-tree.scss are needed to
fix the colors.